Refactored code for base authentication middleware into authentication.js and added useAUthUserSet() help build the needed middleware
