+++ app/graphql/middleware/index.js
@@ -1,14 +1,10 @@
-const makeMiddlewareBundle = require('./makeMiddlewareBundle');
-const requireVerifiedUser = require('./requireVerifiedUser');
-const setAuthTokenUser = require('./setAuthTokenUser');
 const setRoomContext = require('./setRoomContext');
 const setMemberContext = require('./setMemberContext');
+const { useAuthUserSet } = require('./authentication');
 const { usePermissionContextSet } = require('./permissions');
 
 module.exports = {
-    makeMiddlewareBundle,
-    requireVerifiedUser,
-    setAuthTokenUser,
+    useAuthUserSet,
     setRoomContext,
     setMemberContext,
     usePermissionContextSet
